The search engine component now interacts with the Bing Web Search API using a proxy
The config-user now also has the apigateway:PATCH capability. The initialization of the API Gateway has been refactored to include a 'GET /bing' route, which has been integrated with the initialization of the logging endpoint. The API gateway endpoint is then 'exported' by adding it to the environment utilized by the Angular application. In production, the search-engine-body replaces the default dev endpoint by utilizing this environment variable.