Other search engine fixes
Now, the search engine's body components do not trigger additional queries upon pagination exploration if the estimated matches are lower than the SEARCH_AMOUNT factor for each service. Additionally, the query is sent to the APIs in a safer, URL-encoded form. This form is then processed and stored in the final output data.