The Skeleton now (tries to) reinitialize the stepper on the basis of previous history
When a worker rejoins a task after closing their browser tab, the Skeleton attempts to reinitialize the status of each questionnaire and document form using the previously uploaded data records stored on DynamoDB. The ACL now also includes a "position_current" attribute, representing the index of the last element seen by the worker. The Skeleton then downloads the previous data records of the workers, and the most recent one is determined for each element of the assigned HIT. A new DataRecord class is added to parse these downloaded data records. If previous data exists, each data structure used (timestamps, search engine queries, etc.) is assigned the found values, and each form control is populated with the previous answers.