The text annotator component based on a set of predefined available options can restore the previous annotations from the database.
Most of the work involved using the text highlighting library to properly serialize and deserialize the performed highlights. Additionally, the stringified serialization is now stored in the payloads sent to the database. The ngOnInit method gracefully handles the presence of previous records to parse for the current document. A restoreData method has been implemented for the entire Note class hierarchy.