Fix: If the IP address changes but the Worker ID remains the same and is explicitly passed, it could cause an "overwrite" of the original record.
The first change involved transforming the keyset of the ACL table. Before this fix, only the worker identifier was used as the primary key. Now, the IP address serves as the range key. This means that if the same worker identifier is provided from two different IP addresses, there will be two records in the ACL. This change aims to resolve inconsistencies identified in the FakeNewsME task. Additionally, if different worker identifiers are provided from the same IP address, they are all stored in an 'identifiers_provided' field of the ACL.

To achieve this, the sequence key set by the DynamoDB service now includes the IP address. However, this led to several difficulties during data downloading, necessitating extensive modifications to the download script and ultimately prompting changes in the overall approach. This should be taken into consideration when redownloading data from previously deployed tasks.

Additionally, there is an initial development effort towards fixing the countdown feature.